fish-getopts: command line parser for fish
getopts is designed to process command line arguments
that follow the POSIX Utility Syntax Guidelines.

If no arguments are given, 1 is returned.

* getopts does not read the standard input. Use getopts
to collect options and the standard input to process a
stream of data relevant to your program.

* A double dash, --, marks the end of options. Arguments
after this sequence are placed in the default underscore
key, _.

* The getopts describe here is not equivalent to the
getopts builtin found in other shells. This tool is
only available for fish(1).